main_service/requirements.txt
flask==3.0.3
requests==2.31.0
opentelemetry-sdk==1.24.0
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-instrumentation-flask==0.45b0
opentelemetry-instrumentation-requests==0.45b0